Implemented DemoPressurePad.countTraffic,
Changed implementation of DemoPressurePad.getCongestion to use the super keyword

Implemented an instance in DemoSensor named timeSinceEpoch to represent the simulated passage of time
Properly implemented DemoSensor.getCurrentValue method to actually return data.
Properly implemented the DemoSensor.oneSecond method.

Implemented the DemoSpeedCamera.getThreshold to return an actual value and not a placeholder.
Implemenetd the DemoSpeedCamera.averageSpeed to return an actual value and not a placeholder.

Still need to implement the