Intersection.getConnectedIntersections now returns a copy of the instance variable connections (List<Intersection) so that the original instance variable is not edited by anything outside the class.
Unimplemented testToString.java removed.

Route.getSensors now returns a copy of the instance variable sensors (List<Sensor>) so that the original instance variable is not edited by anything outside the class.

Used the setSignal method in the TrafficLight constructor instead of explicitly defining the Enum TrafficSignal

DemoSensor.oneSecond removed print

Removed the protected constructor as it is not required nor specified in the JavaDoc.

Added more descriptive comments for the getTimedItemManager method's singleton pattern.

Implemented the TimedItemManager.oneSecond method so that it calls the oneSecond method for all registered implementation of the TimedItem interface.